fix: move ChatRetry to Shared/Http, simplify retry structure, harden JsonResponseParser CRLF handling
- Move ChatRetry from Shared/Rag/ to Shared/Http/ with namespace Anela.Heblo.Application.Shared.Http (it retries any HTTP/IO operation, not RAG-specific)
- Simplify retry logic: eliminate nested try-inside-catch; retry call now lives outside any catch so OperationCanceledException propagates naturally without wrapping
- Add using Anela.Heblo.Application.Shared.Http to GenerateLeafletHandler while keeping Shared.Rag for IRagQueryExpander
- Harden JsonResponseParser.StripJsonFences: use TrimEnd() before EndsWith check so closing fences with \r\n endings are matched correctly
